+++ Libraries/Quang.Auth.Entities/Group.cs
@@ -10,29 +10,16 @@ namespace Quang.Auth.Entities
     /// </summary>
     public class Group
     {
-        /// <summary>
-        /// 
-        /// </summary>
-        public long Id { get; set; }
-        /// <summary>
-        /// 
-        /// </summary>
-        public long? ParentId { get; set; }
-        /// <summary>
-        /// 
-        /// </summary>
+        public int Id { get; set; }
+
+        public int? ParentId { get; set; }
+
         public string ParentName { get; set; }
-        /// <summary>
-        /// 
-        /// </summary>
+
         public string Name { get; set; }
-        /// <summary>
-        /// 
-        /// </summary>
+
         public string Description { get; set; }
-        /// <summary>
-        /// 
-        /// </summary>
+
         public int TotalMembers { get; set; }
     }
 }
